# server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_AboloreDev_geritcht-restaurant_internals_dto.AllergenResponse:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  github_com_AboloreDev_geritcht-restaurant_internals_dto.CreateAllergenRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
...
  github_com_AboloreDev_geritcht-restaurant_internals_dto.ForgotPasswordRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
...
  github_com_AboloreDev_geritcht-restaurant_internals_dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
  github_com_AboloreDev_geritcht-restaurant_internals_dto.UpdateAllergenRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
...
  github_com_AboloreDev_geritcht-restaurant_internals_utils.PaginatedResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      meta:
        $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.PaginatedMeta'
      status:
        type: boolean
    type: object
  github_com_AboloreDev_geritcht-restaurant_internals_utils.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      status:
        type: boolean
    type: object
host: https://geritcht-restaurant.onrender.com
info:
  contact:
    email: alabifathiu@yahoo.com
    name: Alabi Fathiu
    url: http://linkedin.com/in/fathiu-alabi
  description: A modern Restaurant API built with Go, Gin, and GORM
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Geritcht Restaurant API
  version: "1.0"
paths:
  /allergens:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of all allergens.
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Allergens retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_dto.AllergenResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
      security:
      - BearerAuth: []
      summary: List allergens
      tags:
      - Allergens
    post:
      consumes:
      - application/json
      description: Create a new allergen. Admin access required.
      parameters:
      - description: Allergen details
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_dto.CreateAllergenRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Allergen created successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_dto.AllergenResponse'
              type: object
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "409":
          description: Allergen already exists
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
      security:
      - BearerAuth: []
      summary: Create an allergen
      tags:
      - Allergens
  /allergens/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an allergen by its ID. Admin access required.
      parameters:
      - description: Allergen ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Allergen deleted successfully
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "400":
          description: Invalid allergen ID
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "404":
          description: Allergen not found
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "409":
          description: Allergen cannot be deleted because it is in use
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
      security:
      - BearerAuth: []
      summary: Delete an allergen
      tags:
      - Allergens
    patch:
      consumes:
      - application/json
      description: Update an existing allergen. Admin access required.
      parameters:
      - description: Allergen ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated allergen details
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_dto.UpdateAllergenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Allergen updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_dto.AllergenResponse'
              type: object
        "400":
          description: Invalid request data or allergen ID
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "404":
          description: Allergen not found
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "409":
          description: Allergen name already exists
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
      security:
      - BearerAuth: []
      summary: Update an allergen
      tags:
      - Allergens
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Send a password reset code to the user's registered email address.
      parameters:
      - description: Forgot password request
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_dto.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password reset code sent successfully
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "429":
          description: Too many requests
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
      summary: Forgot password
      tags:
      - Authentication
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password
      parameters:
      - description: User login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            allOf:
            - $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_dto.AuthResponse'
              type: object
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "429":
          description: Too many requests
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
      summary: User login
      tags:
      - Authentication
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Invalidate refresh token and logout user
      parameters:
      - description: Refresh token to invalidate
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_dto.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Logout successful
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
      summary: User logout
      tags:
      - Authentication
  /auth/password-change:
    post:
      consumes:
      - application/json
      description: Change user password
      parameters:
      - description: Change password data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_dto.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password changed successfully
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "400":
          description: Invalid request data or current password is incorrect
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
      security:
      - BearerAuth: []
      summary: User change password
      tags:
      - Authentication
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Generate a new access token using a valid refresh token
      produces:
      - application/json
      responses:
        "200":
          description: New access token generated
          schema:
            allOf:
            - $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_dto.AuthResponse'
              type: object
        "401":
          description: Invalid or expired refresh token
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
      summary: Refresh access token
      tags:
      - Authentication
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account with email,firstname, lastname and password
      parameters:
      - description: User registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User registered successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_dto.AuthResponse'
              type: object
        "400":
          description: Invalid request data or user already exists
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "429":
          description: Too many requests
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
      summary: Register a new user
      tags:
      - Authentication
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: Reset user password with reset token
      parameters:
      - description: Reset password data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_dto.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password reset successful
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "400":
          description: Invalid request data or token expired
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "429":
          description: Too many requests
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
      summary: User reset password
      tags:
      - Authentication
  /auth/verify-email:
    post:
      consumes:
      - application/json
      description: Verify a user's email address using the verification token or OTP.
      parameters:
      - description: Email verification request
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_dto.VerifyEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Email verified successfully
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "400":
          description: Invalid or expired verification token
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "429":
          description: Too many requests
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
      summary: Verify email
      tags:
      - Authentication
...
  /availability:
    get:
      consumes:
      - application/json
      description: Check available tables for a specified date, time, and party size.
      parameters:
      - description: Reservation date (YYYY-MM-DD)
        in: query
        name: date
        required: true
        type: string
      - description: Reservation time (HH:MM)
        in: query
        name: time
        required: true
        type: string
      - description: Number of guests
        in: query
        name: guests
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Availability retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_dto.TableResponse'
                  type: array
              type: object
        "400":
          description: Invalid request data or time slot
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
      summary: Check table availability
      tags:
      - Reservations
  /cart:
    get:
      description: Retrieve current user's shopping cart with all items
      produces:
      - application/json
      responses:
        "200":
          description: Cart retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_dto.CartResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "404":
          description: Cart not found
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
      security:
      - BearerAuth: []
      summary: Get user's cart
      tags:
      - Cart
    post:
      consumes:
      - application/json
      description: Add a menu item to the current user's shopping cart
      parameters:
      - description: Menu item and quantity to add
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_dto.AddToCartRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Item added to cart successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_dto.CartResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "404":
          description: Menu item not found
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
      security:
      - BearerAuth: []
      summary: Add item to cart
      tags:
      - Cart
...
  /categories/search:
    get:
      consumes:
      - application/json
      description: Search for categories by name or description
      parameters:
      - description: Search query
        in: query
        name: q
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of items per page
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Categories retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
            - properties:
                ' meta':
                  $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.PaginatedMeta'
                data:
                  items:
                    $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_dto.MenuCategoryResponse'
                  type: array
              type: object
        "400":
          description: Invalid search parameters
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "404":
          description: No categories found matching the search criteria
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
      summary: Search categories
      tags:
      - Categories
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"