# server/internals/server/server.go
package server

import (
	"net/http"
	"time"

	_ "github.com/AboloreDev/geritcht-restaurant/docs"
	"github.com/AboloreDev/geritcht-restaurant/internals/config"
	"github.com/AboloreDev/geritcht-restaurant/internals/interfaces"
	"github.com/AboloreDev/geritcht-restaurant/internals/services"
	websockets "github.com/AboloreDev/geritcht-restaurant/internals/web-sockets"
	"github.com/gin-gonic/gin"
	"github.com/rs/zerolog"
	"gorm.io/gorm"

	swaggerFiles "github.com/swaggo/files"
	ginSwagger "github.com/swaggo/gin-swagger"
)

type Server struct {
	cfg                 *config.Config
	log                 zerolog.Logger
	db                  *gorm.DB
	authServices        services.AuthServiceInterface
	redisStore          interfaces.Cacher
	uploadServices      services.UplaodServiceInterface
	categoryServices    services.CategoryServiceInterface
	menuServices        services.MenuServiceInterface
	allergenServices    services.AllergenServiceInterface
	dietaryTagsService  services.DietaryTagsServiceInterface
	userServices        services.UserServiceInterface
	reservationServices services.ReservationServiceInterface
	waitlistService     services.WaitlistServiceInterface
	tableServices       services.TableServiceInterface
	paymentService      services.PaymentServiceInterface
	orderService        services.OrderServiceInterface
	cartServices        services.CartServiceInterface
	hub                 *websockets.Hub
	ingredientService   services.IngredientServiceInterface
	recipesService      services.ReceipesServiceInterface
	inventoryService    services.InventoryServiceInterface
}

func NewServer(
	cfg *config.Config,
	db *gorm.DB,
	log zerolog.Logger,
	authServices services.AuthServiceInterface,
	redisStore interfaces.Cacher,
	uploadServices services.UplaodServiceInterface,
	categoryServices services.CategoryServiceInterface,
	menuServices services.MenuServiceInterface,
	allergenServices services.AllergenServiceInterface,
	dietaryTagsService services.DietaryTagsServiceInterface,
	reservationServices services.ReservationServiceInterface,
	waitlistService services.WaitlistServiceInterface,
	userServices services.UserServiceInterface,
	tableServices services.TableServiceInterface,
	paymentService services.PaymentServiceInterface,
	orderService services.OrderServiceInterface,
	cartServices services.CartServiceInterface,
	hub *websockets.Hub,
	ingredientService services.IngredientServiceInterface,
	recipesService services.ReceipesServiceInterface,
	inventoryService services.InventoryServiceInterface) *Server {
	return &Server{
		cfg:                 cfg,
		log:                 log,
		db:                  db,
		authServices:        authServices,
		redisStore:          redisStore,
		uploadServices:      uploadServices,
		categoryServices:    categoryServices,
		menuServices:        menuServices,
		allergenServices:    allergenServices,
		dietaryTagsService:  dietaryTagsService,
		userServices:        userServices,
		waitlistService:     waitlistService,
		reservationServices: reservationServices,
		tableServices:       tableServices,
		paymentService:      paymentService,
		orderService:        orderService,
		cartServices:        cartServices,
		hub:                 hub,
		ingredientService:   ingredientService,
		recipesService:      recipesService,
		inventoryService:    inventoryService,
	}
}

func (s *Server) SetUpRoutes() *gin.Engine {
	router := gin.New()

	// Static Middlewares
	router.Use(gin.Logger())
	router.Use(gin.Recovery())
	router.Use(s.CORS())

	// ROUTES
	// Health check route
	router.GET("/health", s.HealthCheck)

	// Add documentation routes
	router.GET("/docs/*any", ginSwagger.WrapHandler(swaggerFiles.Handler))

	router.StaticFile("/api-docs", "./docs/rapiddoc.html")

	api := router.Group("/api/v1")

	{
		auth := api.Group("/auth")
		{
			// Auth Routes
			auth.POST("/register", s.RateLimiter(10, time.Minute), s.RegisterUserHandler)
			auth.POST("/login", s.RateLimiter(5, time.Minute), s.LoginUserHandler)
			auth.POST("/logout", s.LogoutHandler)
			auth.POST("/refresh", s.RefreshTokenHandler)
			auth.POST("/forgot-password", s.RateLimiter(3, time.Minute), s.ForgotPasswordHandler)
			auth.POST("/reset-password", s.RateLimiter(5, time.Minute), s.ResetPasswordHandler)
			auth.POST("/verify-email", s.RateLimiter(5, time.Minute), s.VerifyEmailHandler)
			auth.POST("/verify-reset-otp", s.RateLimiter(20, time.Minute), s.VerifyResetOTPHandler)

		}
		protected := api.Group("/")
		protected.Use(s.AuthMiddleware())
		{
			users := protected.Group("/users")
			{
				// User Protected Routes
				users.PATCH("/password-change", s.RateLimiter(20, time.Minute), s.ChangePasswordHandler)
				users.GET("/profile", s.GetUserProfileHandler)
				users.PATCH("/profile", s.RateLimiter(20, time.Minute), s.UpdateUserProfileHandler)
				users.PATCH("/profile/deactivate", s.DeactivateUserHandler)
				users.PATCH("/profile/activate", s.ActivateUserHandler)
				users.GET("/", s.AdminMiddleware(), s.GetAllUserHandler)
				users.GET("/search", s.AdminMiddleware(), s.SearchUserHandler)
			}

			staffs := protected.Group("/staff")
			{
				// Staff Protected Routes
				staffs.GET("/profile", s.StaffMiddleware(), s.GetStaffProfileHandler)
				staffs.PATCH("/profile", s.StaffMiddleware(), s.UpdateStaffProfileHandler)
				staffs.PATCH("/profile/deactivate", s.StaffMiddleware(), s.DeactivateStaffHandler)
				staffs.PATCH("/profile/activate", s.StaffMiddleware(), s.ActivateStaffHandler)
				staffs.GET("/", s.AdminMiddleware(), s.GetAllStaffsHandler)
			}

			category := protected.Group("/categories")
			{
				// Category Protected Routes
				category.POST("/", s.AdminMiddleware(), s.CreateCategoryHandler)
				category.PATCH("/:id", s.AdminMiddleware(), s.UpdateCategoryHandler)
				category.DELETE("/:id", s.AdminMiddleware(), s.DeleteCategory)
				// category.GET("/categories", s.GetCategoriesHandler)
				// category.GET("/categories/:id", s.GetCategory)
			}

			menu := protected.Group("/menu")
			{
				// Menu Protected Routes
				menu.POST("/", s.RateLimiter(20, time.Minute), s.AdminMiddleware(), s.CreateMenuHandler)
				menu.PATCH("/:id", s.RateLimiter(20, time.Minute), s.AdminMiddleware(), s.UpdateMenuHandler)
				// menu.GET("/", s.GetAllMenuHandler)
				// menu.GET("/:id", s.GetMenuHandler)
				menu.DELETE("/:id", s.AdminMiddleware(), s.DeleteMenuHandler)
				menu.PATCH("/:id/toggle", s.AdminMiddleware(), s.ToggleMenuAvailabilityHandler)
				menu.POST("/:id/images", s.AdminMiddleware(), s.UploadMenuImageHandler)
				menu.DELETE("/images/:id", s.AdminMiddleware(), s.DeleteMenuImageHandler)
			}

			allergens := protected.Group("/allergens")
			{
				// Allergens and Dietary Tags Protected Routes
				allergens.POST("/", s.RateLimiter(20, time.Minute), s.AdminMiddleware(), s.CreateAllergenHandler)
				allergens.GET("/", s.AdminMiddleware(), s.GetAllAllergenHandler)
				allergens.PATCH("/:id", s.AdminMiddleware(), s.UpdateAllegenHandler)
				allergens.DELETE("/:id", s.RateLimiter(20, time.Minute), s.AdminMiddleware(), s.DeleteAllergenHandler)

			}

			tags := protected.Group("/tags")
			{
				tags.POST("/", s.RateLimiter(20, time.Minute), s.AdminMiddleware(), s.CreateDietaryTagHandler)
				tags.GET("/", s.AdminMiddleware(), s.GetAllDietaryTagHandler)
				tags.PATCH("/:id", s.RateLimiter(20, time.Minute), s.AdminMiddleware(), s.UpdateDietaryTagHandler)
				tags.DELETE("/:id", s.AdminMiddleware(), s.DeleteDietaryTagHandler)
			}

			table := protected.Group("/tables")
			{
				// Table Protected Routes
				table.POST("/", s.AdminMiddleware(), s.CreateTableHandler)
				// table.GET("/", s.GetAllTablesHandler)
				// table.GET("/:id", s.GetTableHandler)
				table.PATCH("/:id", s.AdminMiddleware(), s.UpdateTableHandler)
				table.DELETE("/:id", s.AdminMiddleware(), s.DeleteTableHandler)
			}

			reservation := protected.Group("/reservations")
			{
				// Reservation Protected Routes
				reservation.POST("/", s.RateLimiter(20, time.Minute), s.CreateReservationHandler)
				reservation.GET("/", s.RoleMiddleware("admin", "staff"), s.GetAllReservationsHandler)
				reservation.GET("/:id/user", s.GetUserReservationHandler)
				reservation.GET("/user", s.GetAllUserReservationsHandler)
				reservation.GET("/today", s.GetTodayReservationHandler)
				reservation.POST("/:id", s.RateLimiter(20, time.Minute), s.CheckInReservationHandler)
				reservation.PATCH("/:id/cancel", s.RateLimiter(10, time.Minute), s.RoleMiddleware("admin", "staff"), s.CancelReservationHandler)
				reservation.GET("/search", s.AdminMiddleware(), s.SearchReservationHandler)
			}

			cart := protected.Group("/cart")
			{
				// Cart Protected Routes
				cart.GET("/", s.GetUserCart)
				cart.POST("/", s.AddToCartHandler)
				cart.PATCH("/:id", s.UpdateCartItemHandler)
				cart.DELETE("/", s.ClearCartHandler)
				cart.DELETE("/:id", s.RemoveCartItemHandler)
			}

			order := protected.Group("/orders")
			{
				// Order Protected Routes
				order.POST("/takeout", s.RateLimiter(20, time.Minute), s.CreateTakeoutOrderHandler)
				order.GET("/takeout/:id", s.GetTakeoutOrderHandler)
				order.GET("/takeout/all", s.GetAllUserTakeoutOrdersHandler)
				order.GET("/all", s.AdminMiddleware(), s.GetAllOrdersHandler)
				order.PATCH("/takeout/:id/cancel", s.RateLimiter(10, time.Minute), s.CancelReservationHandler)
				order.GET("/search", s.AdminMiddleware(), s.SearchOrderHandler)
			}

			payment := protected.Group("/payments")
			{
				// Payment Protected Routes
				payment.POST("/initialize", s.RateLimiter(10, time.Minute), s.InitilaisePaymentHandler)
				payment.GET("/verify/:reference", s.RateLimiter(20, time.Minute), s.VerifyPaymentHandler)
				payment.GET("/history", s.GetAllPaymentHistory)
				payment.GET("/:id", s.GetPaymentDetailsHandler)
				payment.GET("/refund/:id", s.GetRefundDetailsHandler)
				payment.GET("/ref/:reference", s.GetPaymentByReferenceHandler)

			}
			websocket := protected.Group("/ws")
			{
				// Websocket
				websocket.GET("/orders/:id", s.AuthMiddleware(), s.WebSocketHandler)
			}
			ingredient := protected.Group("/ingredients")
			{
				// Ingredient Protected Routes
				ingredient.POST("/", s.RateLimiter(20, time.Minute), s.AdminMiddleware(), s.CreateIngredientHandler)
				ingredient.GET("/", s.AdminMiddleware(), s.GetAllIngredientHandler)
				ingredient.GET("/:id", s.AdminMiddleware(), s.GetIngredientHandler)
				ingredient.PATCH("/:id", s.RateLimiter(20, time.Minute), s.AdminMiddleware(), s.UpdateIngredientHandler)
				ingredient.DELETE("/:id", s.AdminMiddleware(), s.DeleteIngredientHandler)
				ingredient.GET("/low-stock", s.AdminMiddleware(), s.GetLowStockIngredientsHandler)
				ingredient.POST("/limit", s.RateLimiter(20, time.Minute), s.AdminMiddleware(), s.SetThresholdLimitHandler)
				ingredient.GET("/search", s.AdminMiddleware(), s.SearchIngredientHandler)
				ingredient.GET("/check-low-stock", s.AdminMiddleware(), s.CheckLowStockHandler)
				ingredient.GET("/alerts", s.AdminMiddleware(), s.GetInventoryAlertsHandler)
			}
			recipes := protected.Group("/recipes")
			{
				// Recipes Protected Routes
				recipes.POST("/:id", s.RateLimiter(20, time.Minute), s.AdminMiddleware(), s.AddRecipeHandler)
				recipes.GET("/:id", s.AdminMiddleware(), s.GetRecipesHandler)
				recipes.PATCH("/:id", s.RateLimiter(20, time.Minute), s.AdminMiddleware(), s.UpdateRecipeHandler)
				recipes.DELETE("/:id", s.AdminMiddleware(), s.DeleteRecipeHandler)
			}
			waitlist := protected.Group("/waitlist")
			{
				// Waitlist Protected Routes
				waitlist.POST("/", s.RateLimiter(20, time.Minute), s.JoinWaitlistHandler)
				waitlist.GET("/", s.RateLimiter(20, time.Minute), s.GetWaitlistPositionHandler)
			}
		}

		// Public routes
		api.GET("/categories", s.RateLimiter(100, time.Minute), s.GetCategoriesHandler)
		api.GET("/categories/:id", s.RateLimiter(100, time.Minute), s.GetCategory)
		api.GET("/category/search", s.SearchCategoryHandler)
		api.GET("/menu/search", s.SearchMenuHandler)
		api.GET("/menu", s.RateLimiter(100, time.Minute), s.GetAllMenuHandler)
		api.GET("/menu/:id", s.RateLimiter(100, time.Minute), s.GetMenuHandler)
		api.GET("/table", s.RateLimiter(100, time.Minute), s.GetAllTablesHandler)
		api.GET("/table/:id", s.GetTableHandler)
		api.GET("/availability", s.RateLimiter(60, time.Minute), s.CheckAvailabilityHandler)
		api.POST("/payments/webhook", s.WebhookHandler)
	}
	return router
}

func (s *Server) HealthCheck(ctx *gin.Context) {
	ctx.JSON(http.StatusOK, gin.H{
		"status": "ok",
	})
}

func (s *Server) CORS() gin.HandlerFunc {
	return func(ctx *gin.Context) {

		ctx.Header("Access-Control-Allow-Origin", "http://localhost:3000")
		ctx.Header("Access-Control-Allow-Methods", "GET, POST, PUT, PATCH, DELETE, OPTIONS")
		ctx.Header("Access-Control-Allow-Headers", "Content-Type, Authorization")
		ctx.Header("Access-Control-Allow-Credentials", "true")

		if ctx.Request.Method == "OPTIONS" {
			ctx.AbortWithStatus(204)
			return
		}

		ctx.Next()
	}
}